Image regeneration fixed
We now have added the prompt to the regeneration box. Once an image is generated, you can regenerate. Also, when you close the pop-up, it then goes back to the actual content type on the stepper menu that you were working on. It doesn't just go back to the first one.